+++ src/tensor/tensor_scale_d.cc
@@ -32,7 +32,7 @@ const Tensor<double> scale(const Tensor<double> &t, int ndx,
                            const Tensor<double> &v) {
   index d1, d2, d3;
   Tensor<double> output(t.dimensions());
-  ndx = normalize_index(ndx, t.rank());
+  ndx = (int)normalize_index(ndx, t.rank());
   surrounding_dimensions(t.dimensions(), ndx, &d1, &d2, &d3);
   if (d2 != v.size()) {
     std::cerr << "In scale() the dimension " << ndx
